#168 Add Config.isValueDefined and Config.isKeyAndValueDefined methods
Deprecating containsKey since if used for flow control it may
lead to bugs: if a key is defined it is no longer possible to
override it to be not present.

isKeyAndValueDefined will have a similar behaviour to containsKey
but the "" value can be used to override the missing key.

isValuePresent will insist that a value is explicitly set to ""
rather than not defined to reduce the chance of accidentally forgetting
to set a config value.

(cherry picked from commit 8fbf6534c1f365c60b32614dab57a59f1928c4b3)