chore(deps): delete findbugs
* chore(deps): delete findbugs

This change removes the findbugs plugin. Findbugs is superseded by SpotBugs. This is a dead build phase dependency. Spotbugs forked the dead findbugs. Even findbugs says to use spotbugs, which we already do. Before this change we are checking twice, once with a decade out of date version (findbugs).

* fix(deps): update findbugs annotations to spotbugs annotations to v4.9.1

This change updates the findbugs annotations to spotbugs-annotations. Also the version number is updated to latest.

Maven indicates the artifact has been moved as expressed in this change https://mvnrepository.com/artifact/com.google.code.findbugs/jsr305

---------

Co-authored-by: colin-janke <30922222+colin-janke@users.noreply.github.com>