feat: PDASH-3970 add OneToManyCountValue cache index
This index effectively allows multiple counters as part of a single index that are kept up to date with one update. This is an alternative to adding multiple PredicateCountValues that are all updated every time objects are updated.

(cherry picked from commit 00eeb369eba511f43d26ec3fde4b392821880c70)